// src/services/types.ts
export type ErrorObject = {
  success: false;
  status_code: number;
  status_message: string;
};

export interface FailedResponse {
  status: "FAILED";
  data: ErrorObject;
}

export interface SuccessResponse<T> {
  status: "SUCCESS";
  data: T;
}

export type Response<T> = SuccessResponse<T> | FailedResponse;

export type GetRequest<T> = () => Promise<Response<T>>;

// Data Retrieved Types
export interface Genre {
  id: number;
  name: string;
}
export interface GenreList {
  genres: Genre[];
}
export interface Movie {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}
export interface MovieList {
  page: number;
  total_pages: number;
  total_results: number;
  results: Movie[];
}

export interface MovieDetails {
  adult: boolean;
  backdrop_path: string;
  belongs_to_collection: {
    id: number;
    name: string;
    poster_path: string;
    backdrop_path: string;
  };
  budget: number;
  genres: Genre[];
  homepage: string;
  id: number;
  imdb_id: string;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  production_companies: {
    id: number;
    logo_path: string;
    name: string;
    origin_country: string;
  }[];
  production_countries: {
    iso_3166_1: string;
    name: string;
  }[];
  release_date: string;
  revenue: number;
  runtime: number;
  spoken_languages: {
    english_name: string;
    iso_639_1: string;
    name: string;
  }[];
  status: string;
  tagline: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}